Split game.ts: extract game-status and role-bucket into dedicated modules
game.ts exceeded the 400-line hard cap (477 lines). Extract two self-contained
type domains into their own files:

- src/lib/types/game-status.ts — GameStatus enum, StartingGameStatus,
  PlayingGameStatus, FinishedGameStatus, GameWinner, GameStatusState
- src/lib/types/role-bucket.ts — RoleBucketSlot, SimpleRoleBucket,
  AdvancedRoleBucket, RoleBucket, isSimpleRoleBucket

game.ts now imports what it needs from these modules; index.ts exports all
three modules. No external import paths change.
