+++ src/lib/types/index.ts
@@ -1,2 +1,4 @@
 export * from "./game";
+export * from "./game-status";
 export * from "./mode-config";
+export * from "./role-bucket";
